chore: updates from comments
